localservicebookingnew/backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7